# api/openapi.yaml
openapi: 3.1.0
info:
  title: Meeting Agent Local API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:8765/api
security:
  - sessionToken: []
paths:
...
  /capture/start:
    post:
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                targetId: { type: string }
      responses:
        "202": { description: Capture is starting }
        "409": { description: Capture is already active }
  /capture/stop:
    post:
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "202": { description: Capture is stopping }
...
components:
  securitySchemes:
    sessionToken:
      type: apiKey
      in: header
      name: Authorization
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
...